Version 0.9.7
It is highly recommended to update to this version, because of a security vulnerability. 

FIX: SECURITY RELEVANT - TLS was not used if you have chosen this option. (Thanks to Jim Carrier for the bug report.)
ADD: First WordPress MU prototype. Read mu/readme_wpmu.txt for further informations.
FIX: Usernames will be converted to lower case, because usernames are case sensitive in WordPress but not in Active Directory. (Thanks to Robert Nelson for the bug report.)